Remove old switching function
